repair wallet incorrectly marking outputs as spent
