+++ modules/wallet/transactionbuilder.go
@@ -90,19 +90,29 @@ func (tb *transactionBuilder) FundSiacoins(amount types.Currency) error {
 	// Create and fund a parent transaction that will add the correct amount of
 	// siacoins to the transaction.
 	var fund types.Currency
+	// potentialFund tracks the balance of the wallet including outputs that
+	// have been spent in other unconfirmed transactions recently. This is to
+	// provide the user with a more useful error message in the event that they
+	// are overspending.
+	var potentialFund types.Currency
 	parentTxn := types.Transaction{}
+	var spentScoids []types.SiacoinOutputID
 	for scoid, sco := range tb.wallet.siacoinOutputs {
 		// Check that this output has not recently been spent by the wallet.
 		spendHeight := tb.wallet.spentOutputs[types.OutputID(scoid)]
-		if spendHeight > tb.wallet.consensusSetHeight-RespendTimeout {
+		// Prevent an underflow error.
+		allowedHeight := tb.wallet.consensusSetHeight - RespendTimeout
+		if tb.wallet.consensusSetHeight < RespendTimeout {
+			allowedHeight = 0
+		}
+		if spendHeight > allowedHeight {
+			potentialFund = potentialFund.Add(sco.Value)
 			continue
 		}
 		outputUnlockConditions := tb.wallet.keys[sco.UnlockHash].unlockConditions
 		if tb.wallet.consensusSetHeight < outputUnlockConditions.Timelock {
 			continue
 		}
-		// Mark the output as spent.
-		tb.wallet.spentOutputs[types.OutputID(scoid)] = tb.wallet.consensusSetHeight
 
 		// Add a siacoin input for this output.
 		sci := types.SiacoinInput{
@@ -110,12 +120,17 @@ func (tb *transactionBuilder) FundSiacoins(amount types.Currency) error {
 			UnlockConditions: outputUnlockConditions,
 		}
 		parentTxn.SiacoinInputs = append(parentTxn.SiacoinInputs, sci)
+		spentScoids = append(spentScoids, scoid)
 
 		// Add the output to the total fund
 		fund = fund.Add(sco.Value)
+		potentialFund = potentialFund.Add(sco.Value)
 		if fund.Cmp(amount) >= 0 {
 			break
 		}
+	}
+	if potentialFund.Cmp(amount) > 0 && fund.Cmp(amount) < 0 {
+		return modules.ErrPotentialDoubleSpend
 	}
 	if fund.Cmp(amount) < 0 {
 		return modules.ErrLowBalance
@@ -146,7 +161,8 @@ func (tb *transactionBuilder) FundSiacoins(amount types.Currency) error {
 		parentTxn.SiacoinOutputs = append(parentTxn.SiacoinOutputs, refundOutput)
 	}
 
-	// Sign all of the inputs to the parent trancstion.
+	// Sign all of the inputs to the parent trancstion. This is done at the end
+	// so that the outputs aren't marked as spent in the event of an error.
 	for _, sci := range parentTxn.SiacoinInputs {
 		err := addSignatures(&parentTxn, types.FullCoveredFields, sci.UnlockConditions, crypto.Hash(sci.ParentID), tb.wallet.keys[sci.UnlockConditions.UnlockHash()])
 		if err != nil {
@@ -162,6 +178,11 @@ func (tb *transactionBuilder) FundSiacoins(amount types.Currency) error {
 	tb.parents = append(tb.parents, parentTxn)
 	tb.siacoinInputs = append(tb.siacoinInputs, len(tb.transaction.SiacoinInputs))
 	tb.transaction.SiacoinInputs = append(tb.transaction.SiacoinInputs, newInput)
+
+	// Mark all outputs that were spent as spent.
+	for _, scoid := range spentScoids {
+		tb.wallet.spentOutputs[types.OutputID(scoid)] = tb.wallet.consensusSetHeight
+	}
 	return nil
 }
