+++ modules/wallet/wallet_test.go
@@ -168,66 +168,3 @@ func TestNilInputs(t *testing.T) {
 		t.Error(err)
 	}
 }
-
-// TestSendSiacoins probes the SendSiacoins method of the wallet.
-func TestSendSiacoins(t *testing.T) {
-	if testing.Short() {
-		t.SkipNow()
-	}
-
-	// Create a wallet tester.
-	wt, err := createWalletTester("TestSendSiacoins")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer wt.closeWt()
-
-	// Get the initial balance - should be 1 block. The unconfirmed balances
-	// should be 0.
-	confirmedBal, _, _ := wt.wallet.ConfirmedBalance()
-	unconfirmedOut, unconfirmedIn := wt.wallet.UnconfirmedBalance()
-	if confirmedBal.Cmp(types.CalculateCoinbase(1)) != 0 {
-		t.Error("unexpected confirmed balance")
-	}
-	if unconfirmedOut.Cmp(types.ZeroCurrency) != 0 {
-		t.Error("unconfirmed balance should be 0")
-	}
-	if unconfirmedIn.Cmp(types.ZeroCurrency) != 0 {
-		t.Error("unconfirmed balance should be 0")
-	}
-
-	// Send 5000 hastings. The wallet will automatically add a fee. Outgoing
-	// unconfirmed siacoins - incoming unconfirmed siacoins should equal 5000 +
-	// fee.
-	tpoolFee := types.NewCurrency64(10).Mul(types.SiacoinPrecision)
-	_, err = wt.wallet.SendSiacoins(types.NewCurrency64(5000), types.UnlockHash{})
-	if err != nil {
-		t.Fatal(err)
-	}
-	confirmedBal2, _, _ := wt.wallet.ConfirmedBalance()
-	unconfirmedOut2, unconfirmedIn2 := wt.wallet.UnconfirmedBalance()
-	if confirmedBal2.Cmp(confirmedBal) != 0 {
-		t.Error("confirmed balance changed without introduction of blocks")
-	}
-	if unconfirmedOut2.Cmp(unconfirmedIn2.Add(types.NewCurrency64(5000)).Add(tpoolFee)) != 0 {
-		t.Error("sending siacoins appears to be ineffective")
-	}
-
-	// Move the balance into the confirmed set.
-	b, _ := wt.miner.FindBlock()
-	err = wt.cs.AcceptBlock(b)
-	if err != nil {
-		t.Fatal(err)
-	}
-	confirmedBal3, _, _ := wt.wallet.ConfirmedBalance()
-	unconfirmedOut3, unconfirmedIn3 := wt.wallet.UnconfirmedBalance()
-	if confirmedBal3.Cmp(confirmedBal2.Add(types.CalculateCoinbase(2)).Sub(types.NewCurrency64(5000)).Sub(tpoolFee)) != 0 {
-		t.Error("confirmed balance did not adjust to the expected value")
-	}
-	if unconfirmedOut3.Cmp(types.ZeroCurrency) != 0 {
-		t.Error("unconfirmed balance should be 0")
-	}
-	if unconfirmedIn3.Cmp(types.ZeroCurrency) != 0 {
-		t.Error("unconfirmed balance should be 0")
-	}
-}
